latest enhancements in code

+++ src/service/TravelService.jsx
@@ -8,7 +8,7 @@ export const fetchTravels = async (params) => {
 
    
         let response = await axios.get(
-                `http://localhost:8080/admin/travelsPages?name=${params.name}&destination=${params.destination}&duration=${params.duration}&travelers=${params.travelers}&pageNumber=${params.pageNumber}&pageSize=${params.pageSize}&sortDirection=${params.sortDirection}&sortBy=${params.sortBy}`
+                `http://localhost:8080/admin/travelsPages?name=${params.name}&destination=${params.destination}&duration=${params.duration}&travelers=${params.travelers}&type=${params.type}&pageNumber=${params.pageNumber}&pageSize=${params.pageSize}&sortDirection=${params.sortDirection}&sortBy=${params.sortBy}`
         )
         
         return response.data;
@@ -18,34 +18,46 @@ export const deleteTravel = async (id) => {
 
         let token = localStorage.getItem("token");
 
-        let response = await axios.delete(
-                `http://localhost:8080/admin/travels/${id}` , 
-                {
-                        headers: {
-                                Authorization: `Bearer ${token}`
+        if (token) {
+
+                let response = await axios.delete(
+                        `http://localhost:8080/admin/travels/${id}` , 
+                        {
+                                headers: {
+                                        Authorization: `Bearer ${token}`
+                                }
                         }
-                }
-        )
+                )
+        
+        
+                return response;
 
+        }
 
-        return response;
+        
 }
 
 export const fetchTravelById = async (id) => {
 
         let token = localStorage.getItem("token");
 
-        let response = await axios.get(
-                `http://localhost:8080/admin/travels/${id}` , 
-                {
-                        headers: {
-                                Authorization: `Bearer ${token}`
+        if (token) {
+
+                let response = await axios.get(
+                        `http://localhost:8080/admin/travels/${id}` , 
+                        {
+                                headers: {
+                                        Authorization: `Bearer ${token}`
+                                }
                         }
-                }
-        )
+                )
+        
+        
+                return response;
 
+        }
 
-        return response;
+        
 }
 
 
@@ -53,24 +65,30 @@ export const createTravel = async (travel) => {
 
         let token = localStorage.getItem("token");
 
-        try {
+        if ( token) {
 
-        let response = await axios.post(
-                `http://localhost:8080/admin/createTravel`,
-                travel,
-                {
-                        headers: {
-                                Authorization: `Bearer ${token}`
+                try {
+
+                        let response = await axios.post(
+                                `http://localhost:8080/admin/createTravel`,
+                                travel,
+                                {
+                                        headers: {
+                                                Authorization: `Bearer ${token}`
+                                        }
+                                }    
+                        )
+                
+                        return response;
+                
+                        } catch(error) {
+                
+                                console.log(error);
+                
                         }
-                }    
-        )
-
-        return response;
-
-        } catch(error) {
-
-                console.log(error);
 
         }
+
+       
 
 }
